fix: archi

+++ src/engine/collision.c
@@ -6,12 +6,12 @@
 /*   By: achu <achu@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/10 14:41:16 by achu              #+#    #+#             */
-/*   Updated: 2025/04/10 15:24:32 by achu             ###   ########.fr       */
+/*   Updated: 2025/04/10 16:11:33 by achu             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include <math.h>
 #include "engine.h"
-#include "math.h"
 
 // Return True if a vector is inside a rect
 bool	is_point_in_rect(t_vec2 point, t_rect rect)
@@ -36,7 +36,7 @@ t_vec2	ft_penetration(t_rect a, t_rect b)
 {
 	t_vec2	overlap;
 
-	overlap.x = fmin(fabs(a.pos.x + a.size.x - b.pos.x), fabs(b.pos.x + b.size.x - a.pos.x));
-	overlap.y = fmin(fabs(a.pos.y + a.size.y - b.pos.y), fabs(b.pos.y + b.size.y - a.pos.y));
+	overlap.x = fminf(fabsf(a.pos.x + a.size.x - b.pos.x), fabsf(b.pos.x + b.size.x - a.pos.x));
+	overlap.y = fminf(fabsf(a.pos.y + a.size.y - b.pos.y), fabsf(b.pos.y + b.size.y - a.pos.y));
 	return (overlap);
 }
